Refactored Project struture
