+++ backend/src/main/java/de/cae/XYFleet/authentication/digest/LocalVerifier.java
@@ -1,17 +1,16 @@
-package de.cae.XYFleet.authentication;
+package de.cae.XYFleet.authentication.digest;
 
 import de.cae.XYFleet.Main;
 import org.jooq.DSLContext;
 import org.jooq.codegen.XYFleet.tables.records.UsersRecord;
 import org.restlet.resource.ResourceException;
-import org.restlet.security.*;
 
 import static org.jooq.codegen.XYFleet.Tables.USERS;
 
 /**
  * searches for the Password of the given User for Password Verification
  */
-public class LDAPLocalVerifier extends LocalVerifier {
+public class LocalVerifier extends org.restlet.security.LocalVerifier {
     @Override
     public char[] getLocalSecret(String s) {
         DSLContext dslContext = Main.getDSLContext();
